Added command to sim
I wanted to make sure that sending emails via a management command and
cron job was viable given the suspected load on the system and I wanted
to do this using real data.  So, instead of messing with fixtures and
writing a test, I just made a management command to test this.